fix(rm/polyfill): handle ENOENT in rmdir more appropriately
an ENOENT from rmdir should follow the same rules as every
other error handler, with one major exception. in windows,
if the path passed to rmdir is not a directory an ENOENT is
thrown instead of the ENOTDIR present in other platforms.
to determine if we should ignore the ENOENT or throw our own
ENOTDIR, we use an lstat to see if the path still exists